Refactor server setup: move Express app initialization to app.js, streamline middleware usage, and enhance graceful shutdown handling.
