+++ server/index.js
@@ -1,42 +1,24 @@
-import express from "express";
-import cors from "cors";
 import dotenv from "dotenv";
-import userRoutes from "./routes/userRoutes.js";
-import postRoutes from "./routes/postRoutes.js";
-import { errorHandler } from "./middleware/errorHandler.js";
+import app from "./app.js";
+import prisma from "./config/prisma.js";
 
 dotenv.config();
-const app = express();
+
 const PORT = process.env.PORT || 3001;
 
-// Middleware
-app.use(cors());
-app.use(express.json());
-
-// Health check
-app.get("/", (req, res) => {
-  res.json({ message: "Server is running!" });
-});
-
-// Routes
-app.use("/api/users", userRoutes);
-app.use("/api/posts", postRoutes);
-
-// Global error handler
-app.use(errorHandler);
-
 // Start server
-app.listen(PORT, () => {
+const server = app.listen(PORT, () => {
   console.log(`Server is running on http://localhost:${PORT}`);
 });
 
 // Graceful shutdown
-process.on("SIGINT", async () => {
+const shutdown = async () => {
+  console.log("Shutting down gracefully...");
   await prisma.$disconnect();
-  process.exit(0);
-});
+  server.close(() => {
+    process.exit(0);
+  });
+};
 
-process.on("SIGTERM", async () => {
-  await prisma.$disconnect();
-  process.exit(0);
-});
+process.on("SIGTERM", shutdown);
+process.on("SIGINT", shutdown);
